pal: use floating point operation for precise calculation
use floating operation for precise calculation of ftm parameters.

+++ device/src/SpeakerProtection.cpp
@@ -2878,8 +2878,8 @@ int32_t SpeakerProtection::getFTMParameter(void **param)
             if (exFtm_ret[1].status == 4 && ftm_ret[1].status == 4)
                 spkr2_status = 1;
             resString << "SpkrParamStatus: " << spkr1_status <<", "<< spkr2_status
-                    << "; Rdc: " << ((ftm_ret[0].ftm_rDC_q24)/(1<<24)) << ", "
-                    << ((ftm_ret[1].ftm_rDC_q24)/(1<<24)) << "; Temp: "
+                    << "; Rdc: " << (((float)ftm_ret[0].ftm_rDC_q24)/(1<<24)) << ", "
+                    << (((float)ftm_ret[1].ftm_rDC_q24)/(1<<24)) << "; Temp: "
                     << ((ftm_ret[0].ftm_temp_q22)/(1<<22)) << ", "
                     << ((ftm_ret[1].ftm_temp_q22)/(1<<22)) <<"; Res: "
                     << ((exFtm_ret[0].ftm_Re_q24)/(1<<24)) << ", "
@@ -2890,8 +2890,8 @@ int32_t SpeakerProtection::getFTMParameter(void **param)
                     << ((exFtm_ret[1].ftm_Rms_q24)/(1<<24)) << "; Kms: "
                     << ((exFtm_ret[0].ftm_Kms_q24)/(1<<24)) << ", "
                     << ((exFtm_ret[1].ftm_Kms_q24)/(1<<24)) << "; Fres: "
-                    << ((exFtm_ret[0].ftm_Fres_q20)/(1<<20)) << ", "
-                    << ((exFtm_ret[1].ftm_Fres_q20)/(1<<20)) << "; Qms: "
+                    << (((float)exFtm_ret[0].ftm_Fres_q20)/(1<<20)) << ", "
+                    << (((float)exFtm_ret[1].ftm_Fres_q20)/(1<<20)) << "; Qms: "
                     << ((exFtm_ret[0].ftm_Qms_q24)/(1<<24)) << ", "
                     << ((exFtm_ret[1].ftm_Qms_q24)/(1<<24));
             break;
